updated education and employment controllers - added 400 error handler

+++ controllers/employment.js
@@ -1,35 +1,31 @@
 const mongoose = require("mongoose");
 const Employment = require("../models/employment/Employment.js");
 const EmploymentItem = require("../models/employment/EmploymentItem.js");
-// const User = require("../models/User.js");
+const Http404Error = require("../utils/errorHandling/http404Error");
+const Http400Error = require("../utils/errorHandling/http400Error.js");
 
-const getEmployment = (req, res) => {
-  // let employmentId;
-  // if (req.params.id) {
-  //   employmentId = req.params.id;
-  // } else {
-  //   User.findById(req.auth._id, (err, user) => {
-  //     employmentId = user.employment;
-  //   });
-  // }
-
+const getEmployment = async (req, res, next) => {
   try {
-    Employment.find({ user: req.auth._id })
-      .populate(["employmentList"])
-      .exec((err, employment) => {
-        if (err) {
-          employment = new Employment({});
-          employment.save();
-        }
+    const employment = await Employment.findOne({ user: req.auth._id })
+      .populate("employmentList")
+      .exec();
 
-        res.status(200).json(employment[0]);
-      });
+    if (!employment) {
+      employment = new Employment({ user: req.auth._id }).exec();
+      await employment.save();
+    }
+
+    res.status(200).json(employment);
   } catch (error) {
-    res.status(404).json({ message: error.message });
+    next(error);
   }
 };
 
-const createEmploymentItem = async (req, res) => {
+const createEmploymentItem = async (req, res, next) => {
+  if (!req.body.data) {
+    throw new Http400Error("No information was provided");
+  }
+
   const employmentItem = new EmploymentItem({
     ...req.body.data,
     employment: req.params.id,
@@ -38,45 +34,60 @@ const createEmploymentItem = async (req, res) => {
   try {
     await employmentItem.save();
 
-    Employment.findById(employmentItem.employment)
+    const employment = await Employment.findById(employmentItem.employment)
       .populate(["employmentList"])
-      .exec((err, employment) => {
-        employment.employmentList.push(employmentItem);
-        employment.save();
+      .exec();
 
-        res.status(200).json(employment);
-      });
+    if (!employment) {
+      throw new Http404Error("Employment data not found");
+    }
+
+    employment.employmentList.push(employmentItem);
+    await employment.save();
+
+    res.status(200).json(employment);
   } catch (error) {
-    res.status(404).json({ message: error.message });
+    next(error);
   }
 };
 
-const updateEmploymentItem = async (req, res) => {
-  let employmentItem = req.body.data;
+const updateEmploymentItem = async (req, res, next) => {
+  if (!req.body.data) {
+    throw new Http400Error("No information was provided");
+  }
+
+  const employmentItem = req.body.data;
   const id = req.body.data._id;
 
-  if (!mongoose.Types.ObjectId.isValid(id))
-    return res.status(404).send("No item with that id");
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    throw new Http404Error("Employment item not found");
+  }
 
-  employmentItem = await EmploymentItem.findByIdAndUpdate(
+  await EmploymentItem.findByIdAndUpdate(
     id,
     { ...employmentItem, id },
     { new: true }
   );
 
   try {
-    Employment.findById(employmentItem.employment)
-      .populate(["employmentList"])
-      .exec((err, employment) => {
-        res.status(201).json(employment);
-      });
+    const employment = await Employment.findById(employmentItem.employment)
+      .populate("employmentList")
+      .exec();
+
+    if (!employment) {
+      throw new Http404Error("Employment data not found");
+    }
+
+    res.status(201).json(employment);
   } catch (error) {
-    res.status(404).json({ message: error.message });
+    next(error);
   }
 };
 
-const bulkUpdateEmploymentItems = async (req, res) => {
-  if (!req.body.data) return;
+const bulkUpdateEmploymentItems = async (req, res, next) => {
+  if (!req.body.data) {
+    throw new Http400Error("No information was provided");
+  }
 
   const employmentItems = req.body.data;
 
@@ -84,40 +95,53 @@ const bulkUpdateEmploymentItems = async (req, res) => {
     const id = item._id;
 
     if (!mongoose.Types.ObjectId.isValid(id)) {
-      return res.status(404).send("No item with that id");
+      throw new Http404Error("Employment item not found");
     }
 
     await EmploymentItem.findByIdAndUpdate(id, { ...item, id }, { new: true });
   });
 
   try {
-    Employment.findById(req.params.id)
+    const employment = await Employment.findById(req.params.id)
       .populate("employmentList")
-      .exec((err, employment) => {
-        res.status(201).json(employment);
-      });
+      .exec();
+
+    if (!employment) {
+      throw new Http404Error("Employment data not found");
+    }
+
+    res.status(201).json(employment);
   } catch (error) {
-    res.status(404).json({ message: error.message });
+    next(error);
   }
 };
 
-const deleteEmploymentItem = async (req, res) => {
-  let employmentItem = req.body.data;
-  let employmentId = employmentItem.employment;
+const deleteEmploymentItem = async (req, res, next) => {
+  if (!req.body.data) {
+    throw new Http400Error("No information was provided");
+  }
 
-  if (!mongoose.Types.ObjectId.isValid(employmentItem._id))
-    return res.status(404).send("No item with that id");
+  const employmentItem = req.body.data;
+  const employmentId = employmentItem.employment;
+
+  if (!mongoose.Types.ObjectId.isValid(employmentItem._id)) {
+    throw new Http404Error("Employment item not found");
+  }
 
   await EmploymentItem.findByIdAndRemove(employmentItem._id);
 
   try {
-    Employment.findById(employmentId)
-      .populate(["employmentList"])
-      .exec((err, employment) => {
-        res.status(200).json(employment);
-      });
+    const employment = await Employment.findById(employmentId)
+      .populate("employmentList")
+      .exec();
+
+    if (!employment) {
+      throw new Http404Error("Employment data not found");
+    }
+
+    res.status(200).json(employment);
   } catch (error) {
-    res.status(404).json({ message: error.message });
+    next(error);
   }
 };
